Moved Platform rendering completely to Platforms container from Teams
